+++ packages/uilib/src/lib/components/comp-board/comp-board.svelte
@@ -174,6 +174,7 @@
 			<CompColumn
 				index={index}
 				on:linkclick
+				on:createnote
 			/>
 		</li>
 	{/each}